moved jobs into separate subdir and adjusted some small stylings
